fix parsing of git blame
Previously, we were using the line number from the commit, instead of the line number from the current status of the file.
When a file changes over multiple commits, this causes the line list to be wrong, and i18n-calypso will ignore some changes.
Luckily, the 3rd column has the line number we need